requirements.txt
discord.py==1.3.4
boto3==1.11.9
sqlalchemy==1.3.0
mysqlclient==1.4.6
s3fs==0.2.0
pytest==5.0.0
cryptography==2.7
python-dotenv==0.14.0
markovify==0.7.1
numpy==1.16.4
nltk==3.3
scikit-learn==0.19.2
scipy==1.1.0
requests==2.22.0